pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-for-or"
version = "0.1.0"
description = "LLM Agents for Operations Research: Evaluation Benchmarks and Training"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ruicheng Ao", email = "your-email@mit.edu"}
]
keywords = [
    "operations research",
    "large language models",
    "reinforcement learning",
    "optimization",
    "benchmarking"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "gymnasium>=0.29.0",
    "gurobipy>=10.0.0",
    "pyomo>=6.6.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
]
api = [
    "openai>=1.0.0",
    "anthropic>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/LLM-for-OR"
Documentation = "https://github.com/YOUR_USERNAME/LLM-for-OR/tree/main/docs"
Repository = "https://github.com/YOUR_USERNAME/LLM-for-OR"
Issues = "https://github.com/YOUR_USERNAME/LLM-for-OR/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | build
    | dist
    | tmp
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [".git", ".mypy_cache", ".pytest_cache", ".venv", "build", "dist", "tmp"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
